Fix always_on_top and saving/reading settings
Update set_setting to convert value to string since it will be written
to a file.

Change on_top to use 1 and 0 instead of True and False, as these can
easily be converted from string to a "truthy" value, unlike "True" and
"False" without explicitly checking for the string.  Change if
statements to check for truthiness instead of matching explicit bools.

Update read settings code to use defaults if unable to read valid
settings from file.